Cleaned the Camera class. Added two isometric related methods. Changed the Level class so it won't need an reference of the Engine. Entity and Sprite now uses floats instead of integers in the render method.

+++ Greywater/src/com/teamsweepy/greywater/entities/components/Sprite.java
@@ -80,7 +80,7 @@ public class Sprite {
 	 * @param x - x co-ordinate to render at
 	 * @param y - y co-ordinate to render at
 	 */
-	public void render(SpriteBatch g, int x, int y) {
+	public void render(SpriteBatch g, float x, float y) {
 		if (playMode == STILL_IMAGE || playMode == HALTED_PLAYING)
 			g.draw(sprite, x, y);
 		else
